[#noissue] Move AgentParam from web to commons-server
